+++ extras/ci/test.sh
@@ -1,4 +1,10 @@
 #!/bin/sh -ex
 
-"$(dirname "$0")/build.sh"
+export CC="$_CC"
+export CXX="$_CXX"
+
+[ -n "$SANITIZE" ] && export CXXFLAGS="-fsanitize=$SANITIZE"
+
+cmake -DCMAKE_BUILD_TYPE=Debug .
+cmake --build .
 ctest --output-on-failure .