+++ CHANGELOG.md
@@ -6,6 +6,7 @@ HEAD
 
 * CMake: don't build tests when imported in another project
 * CMake: made project arch-independent
+* Visual Studio: fixed error C2766 with flag `/Zc:__cplusplus` (issue #1250)
 
 v6.15.1 (2020-04-08)
 -------