+++ CHANGELOG.md
@@ -8,6 +8,7 @@ HEAD
 * CMake: made project arch-independent
 * Visual Studio: fixed error C2766 with flag `/Zc:__cplusplus` (issue #1250)
 * Added support for `JsonDocument` to `copyArray()` (issue #1255)
+* Added support for `enum`s in `as<T>()` and `is<T>()`  (issue #1256)
 
 v6.15.1 (2020-04-08)
 -------