+++ CHANGELOG.md
@@ -11,6 +11,7 @@ HEAD
 * Added support for `enum`s in `as<T>()` and `is<T>()`  (issue #1256)
 * Added `JsonVariant` as an input type for `deserializeXxx()`  
   For example, you can do: `deserializeJson(doc2, doc1["payload"])`
+* Break the build if using 64-bit integers with ARDUINOJSON_USE_LONG_LONG==0
 
 v6.15.1 (2020-04-08)
 -------